refactor: clear the last new-code smells from the file-split residue
The file-splits left 7 issues SonarCloud's re-scan surfaced: 6 backward-compat
re-exports (X as X) that S1128 does not recognize as used, and one extracted
method that no longer uses self. Repoint the two real callers of the used
re-exports (lab.py, a curated-variants test) at the source module and drop all
six re-export lines; make _write_image_node_content_override static. ruff clean,
imports resolve, targeted suites pass.
